`[✓] ref(ui/our/text): divide CircleTextView, build function by class`
**PREVIOUSLY:**
- CircleTextView set the background and text alignment and controlled sizing.


**NOW:**
- A Shape enum maintains the list of avialable view shape.

- BaseText acts as a base class for the app's text views. It ensures center alignment and allows choosing shape from the enum. It also acts as a host to any additional functionality we'll want to add to text views. and already simplifies setting the background color.

- EllipseTextView extends BaseText and simply specifies the shape as an ellipse.

- CircleTextVIew extends ellipse and ensures the width and height are equal (i.e, that it's a circle rather than an ellipse), as well as supplying a setRadius method.

**BECAUSE:**
- Hierarchaical structure divides unrelated functions and makes for clearer code.

- Distinct funcitonalities that we ight want to use in rheir own right are now available without the extra constraints (e.g extra TextView functionality, non-circular ellipse chape).

NOTES:
We might want to start using BaseText instead of TextView in all of our ui. Then we'll very easily be able to add extra functionality as a baseline.

—

COMPILES: Due to the way I am commiting this, you'll have to remove the OVAL_RECT line from the Shape class, and replace the comma with a semicolon in the line above it. Other than that, yes.
RUNS: Yes.
WORKS: Yes
BUGS: None