[✓] fix(form): fix tab data issue and remove redundancies
Added the important line: setOffPageViewLimits or whatever
but then it didn't work, but that was because everything was handled by the TabLayout and the viewpager was only there.
now everything is created dynamically through viewpager, it saves the data, and you're able to swipe through views. the code is also much cleaner.

also went back to form launcher being the first activity

brought to you