`[✓] bug(acts/notes): force add note buttons to show up`
**BUG:**
The camera note button became a text item in the three dots menu, instead of showing up as an action icon.

**CAUSE:**
By default, Android allows only two action icons to show up.

**FIX:**
in the menu items, specify showAsAction="always" (in position of the previous "ifRoom"), which ignores the button count and forces the icon to show up.