`fix(scouts): changed some text views to buttons to fix a bug`
Text views in various parts of the form, that really should've been buttons all along, were causing problems when getting their backgrounds colored, so I just changed them to buttons.

The bug this caused was that it would crash the application if one had tried to turn on their hotspot after immediately after submitting a form and returning to the form submit page. This fixed it.